move customNodes file
