+++ fresh.gen.ts
@@ -21,12 +21,13 @@ import * as $15 from "./routes/service/[service]/component/[component]/[operatio
 import * as $16 from "./routes/service/[service]/component/[component]/[operationType]/op/[operationName]/pipeline/[id]/pause.tsx";
 import * as $17 from "./routes/service/[service]/component/[component]/[operationType]/op/[operationName]/schema.tsx";
 import * as $18 from "./routes/ws/service-map.tsx";
-import * as $$0 from "./islands/notifications.tsx";
-import * as $$1 from "./islands/opModal.tsx";
-import * as $$2 from "./islands/peek.tsx";
-import * as $$3 from "./islands/pipeline.tsx";
-import * as $$4 from "./islands/pipelines.tsx";
-import * as $$5 from "./islands/serviceMap.tsx";
+import * as $$0 from "./islands/customNodes.tsx";
+import * as $$1 from "./islands/notifications.tsx";
+import * as $$2 from "./islands/opModal.tsx";
+import * as $$3 from "./islands/peek.tsx";
+import * as $$4 from "./islands/pipeline.tsx";
+import * as $$5 from "./islands/pipelines.tsx";
+import * as $$6 from "./islands/serviceMap.tsx";
 
 const manifest = {
   routes: {
@@ -57,12 +58,13 @@ const manifest = {
     "./routes/ws/service-map.tsx": $18,
   },
   islands: {
-    "./islands/notifications.tsx": $$0,
-    "./islands/opModal.tsx": $$1,
-    "./islands/peek.tsx": $$2,
-    "./islands/pipeline.tsx": $$3,
-    "./islands/pipelines.tsx": $$4,
-    "./islands/serviceMap.tsx": $$5,
+    "./islands/customNodes.tsx": $$0,
+    "./islands/notifications.tsx": $$1,
+    "./islands/opModal.tsx": $$2,
+    "./islands/peek.tsx": $$3,
+    "./islands/pipeline.tsx": $$4,
+    "./islands/pipelines.tsx": $$5,
+    "./islands/serviceMap.tsx": $$6,
   },
   baseUrl: import.meta.url,
 };
